Found a critical bug in sgrun : Pipe Deadlock, added a workaround
The pipes have 64kb(usually) buffers, if they fill up, OS suspends the
process until pipes are drained.
This causes pipe deadlock.
currently the sgrun doesnt drain pipes, and tbh I dont like the
structure of runner right now, I will be adding a concurrency library to
drocc later probably to fix this

+++ core/src/core/string_interner.h
@@ -20,7 +20,7 @@ typedef struct InternEntry {
   /// first char of string for faster comparisons
   uint64_t char1 : 8;
   uint64_t is_full : 1;
-  uint32_t intern_id : 32;
+  uint64_t intern_id : 32;
 } InternEntry;
 
 typedef vec(InternEntry) InternEntryVec;
